+++ apps/backend/src/app/node/behaviors/index.ts
@@ -2,5 +2,7 @@ export * from "./node-behavior.base";
 export * from "./node-behavior.code";
 export * from "./node-behavior.entity";
 export * from "./node-behavior.function";
-export * from "./node-behavior.parameter";
+export * from "./node-behavior.parameter-input";
+export * from "./node-behavior.parameter-output";
 export * from "./node-behavior.reference";
+export * from "./node-behavior.variable";