+++ apps/backend/src/app/node/behaviors/node-behavior.variable.ts
@@ -3,14 +3,14 @@ import { NodeBehaviorVariableDto } from "~/lib/common/app/node/dtos/behaviors";
 import { NodeBehaviorType } from "~/lib/common/app/node/dtos/behaviors/node-behavior.type";
 import { NodeIoValue } from "~/lib/common/app/node/io";
 
-import { NodeBehaviorParameterBase } from "./node-behavior.parameter-base";
+import { NodeBehaviorBase } from "./node-behavior.base";
 
 /**
  * Node behavior of `VARIABLE` type
  */
 @Entity({ discriminatorValue: NodeBehaviorType.VARIABLE })
 export class NodeBehaviorVariable
-	extends NodeBehaviorParameterBase<NodeBehaviorType.VARIABLE>
+	extends NodeBehaviorBase<NodeBehaviorType.VARIABLE>
 	implements NodeBehaviorVariableDto
 {
 	/** @inheritDoc */