+++ apps/backend/src/app/node/behaviors/parameters/index.ts
@@ -1,4 +1,0 @@
-export * from "./node-behavior.parameter-base";
-export * from "./node-behavior.parameter-input";
-export * from "./node-behavior.parameter-output";
-export * from "./node-behavior.variable";