+++ libs/common/src/app/node/dtos/node.update.dto.ts
@@ -3,9 +3,15 @@ import { Expose, Type as TypeTransformer } from "class-transformer";
 import { IsOptional, ValidateNested } from "class-validator";
 
 import {
+	NODE_BEHAVIOR_DISCRIMINATOR_KEY,
+	NODE_BEHAVIOR_UPDATE_DTOS,
+	NodeBehaviorBaseDto,
+	NodeBehaviorUpdateDto
+} from "./behaviors";
+import {
+	NODE_KIND_DISCRIMINATOR_KEY,
 	NODE_KIND_UPDATE_DTOS,
 	NodeKindBaseDto,
-	NodeKindDiscriminatorKey,
 	NodeKindUpdateDto
 } from "./kind";
 import { NodeCreateDto } from "./node.create.dto";
@@ -19,9 +25,21 @@ import { NodeCreateDto } from "./node.create.dto";
 export class NodeUpdateDto extends PartialType(OmitType(NodeCreateDto, ["behavior", "kind"])) {
 	@Expose()
 	@IsOptional()
+	@TypeTransformer(() => NodeBehaviorBaseDto, {
+		discriminator: {
+			property: NODE_BEHAVIOR_DISCRIMINATOR_KEY,
+			subTypes: NODE_BEHAVIOR_UPDATE_DTOS.slice()
+		},
+		keepDiscriminatorProperty: true
+	})
+	@ValidateNested()
+	public readonly behavior?: NodeBehaviorUpdateDto;
+
+	@Expose()
+	@IsOptional()
 	@TypeTransformer(() => NodeKindBaseDto, {
 		discriminator: {
-			property: "type" satisfies NodeKindDiscriminatorKey,
+			property: NODE_KIND_DISCRIMINATOR_KEY,
 			subTypes: NODE_KIND_UPDATE_DTOS.slice()
 		},
 		keepDiscriminatorProperty: true